refactor: improve logging stategy
